+++ src/features/manageProfileSettings/components/ProfileSettingsEditor/BaseComponent.test.tsx
@@ -11,29 +11,34 @@ import { UnusedComponent } from './UnusedComponent';
 
 describe('BaseComponent', () => {
   const name = 'test-name';
-  const mockComponent = {
-    id: 'test:component',
-    name,
+  const makeComponent = (mandatory: boolean) => {
+    return {
+      id: 'test:component',
+      name,
+      mandatory,
+    };
   };
 
-  const renderComponent = (node: ReactNode) => {
+  const renderComponent = (node: ReactNode, component: ProfileSettingComponent) => {
     return render(
       <MemoryRouter>
         <DndContext>
-          <SortableContext items={[mockComponent]}>{node}</SortableContext>
+          <SortableContext items={[component]}>{node}</SortableContext>
         </DndContext>
       </MemoryRouter>,
     );
   };
 
   it('renders UnusedComponent', () => {
-    renderComponent(<UnusedComponent component={mockComponent} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<UnusedComponent component={mockComponent} />, mockComponent);
 
     expect(screen.getByText(name)).toBeInTheDocument();
   });
 
   it('renders SelectedComponent with both nudge buttons', () => {
-    renderComponent(<SelectedComponent component={mockComponent} size={3} index={2} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<SelectedComponent component={mockComponent} size={3} index={2} />, mockComponent);
 
     expect(screen.getByText('2. ' + name)).toBeInTheDocument();
     expect(screen.getByTestId('nudge-up')).toBeInTheDocument();
@@ -41,7 +46,8 @@ describe('BaseComponent', () => {
   });
 
   it('renders SelectedComponent at top of list with only nudge down button', () => {
-    renderComponent(<SelectedComponent component={mockComponent} size={3} index={1} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<SelectedComponent component={mockComponent} size={3} index={1} />, mockComponent);
 
     expect(screen.getByText('1. ' + name)).toBeInTheDocument();
     expect(screen.queryByTestId('nudge-up')).not.toBeInTheDocument();
@@ -49,21 +55,31 @@ describe('BaseComponent', () => {
   });
 
   it('renders SelectedComponent at bottom of list with only nudge up button', () => {
-    render(<SelectedComponent component={mockComponent} size={3} index={3} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<SelectedComponent component={mockComponent} size={3} index={3} />, mockComponent);
 
     expect(screen.getByText('3. ' + name)).toBeInTheDocument();
     expect(screen.getByTestId('nudge-up')).toBeInTheDocument();
     expect(screen.queryByTestId('nudge-down')).not.toBeInTheDocument();
   });
 
+  it('renders mandatory SelectedComponent with appropriate text', () => {
+    const mockComponent = makeComponent(true);
+    renderComponent(<SelectedComponent component={mockComponent} size={1} index={1} />, mockComponent);
+
+    expect(screen.getByText('ld.requiredAnnotation')).toBeInTheDocument();
+  });
+
   it('renders DraggingComponent', () => {
-    renderComponent(<DraggingComponent component={mockComponent} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<DraggingComponent component={mockComponent} />, mockComponent);
 
     expect(screen.getByText(name)).toBeInTheDocument();
   });
 
   it('renders a SelectedComponent with a context menu when the button is clicked', () => {
-    render(<SelectedComponent component={mockComponent} size={1} index={1} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<SelectedComponent component={mockComponent} size={1} index={1} />, mockComponent);
 
     expect(screen.getByTestId('activate-menu')).toBeInTheDocument();
     fireEvent.click(screen.getByTestId('activate-menu'));
@@ -73,8 +89,21 @@ describe('BaseComponent', () => {
     });
   });
 
+  it('renders a mandatory SelectedComponent with a context menu with a warning when the button is clicked', () => {
+    const mockComponent = makeComponent(true);
+    renderComponent(<SelectedComponent component={mockComponent} size={1} index={1} />, mockComponent);
+
+    expect(screen.getByTestId('activate-menu')).toBeInTheDocument();
+    fireEvent.click(screen.getByTestId('activate-menu'));
+
+    waitFor(() => {
+      expect(screen.queryByTestId('move-action')).not.toBeInTheDocument();
+    });
+  });
+
   it('renders an UnusedComponent with a context menu when the button is clicked', () => {
-    render(<UnusedComponent component={mockComponent} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<UnusedComponent component={mockComponent} />, mockComponent);
 
     expect(screen.getByTestId('activate-menu')).toBeInTheDocument();
     fireEvent.click(screen.getByTestId('activate-menu'));
@@ -85,7 +114,8 @@ describe('BaseComponent', () => {
   });
 
   it('toggle off a context menu', () => {
-    render(<UnusedComponent component={mockComponent} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<UnusedComponent component={mockComponent} />, mockComponent);
 
     fireEvent.click(screen.getByTestId('activate-menu'));
 
@@ -101,7 +131,8 @@ describe('BaseComponent', () => {
   });
 
   it('dismiss a context menu by clicking elsewhere', () => {
-    render(<UnusedComponent component={mockComponent} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<UnusedComponent component={mockComponent} />, mockComponent);
 
     fireEvent.click(screen.getByTestId('activate-menu'));
 
@@ -117,7 +148,8 @@ describe('BaseComponent', () => {
   });
 
   it('dismiss a context menu by typing escape', () => {
-    render(<UnusedComponent component={mockComponent} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<UnusedComponent component={mockComponent} />, mockComponent);
 
     fireEvent.click(screen.getByTestId('activate-menu'));
 
@@ -133,7 +165,8 @@ describe('BaseComponent', () => {
   });
 
   it('dismiss a context menu by changing focus', () => {
-    render(<SelectedComponent component={mockComponent} size={2} index={1} />);
+    const mockComponent = makeComponent(false);
+    renderComponent(<SelectedComponent component={mockComponent} size={2} index={1} />, mockComponent);
 
     screen.getByTestId('activate-menu').focus();
     fireEvent.click(screen.getByTestId('activate-menu'));